split off disk drives stuff into disks.[hc]
Add functions init_disks() and exit_disks() and use them.
Add defines for number of disks and path length of disks.
